+++ packages/transport-circuit-relay-v2/src/utils.ts
@@ -1,5 +1,6 @@
+import { CODE_P2P_CIRCUIT } from '@multiformats/multiaddr'
 import { P2P } from '@multiformats/multiaddr-matcher'
-import { fmt, literal, and } from '@multiformats/multiaddr-matcher/utils'
+import { fmt, code, and } from '@multiformats/multiaddr-matcher/utils'
 import { anySignal } from 'any-signal'
 import { CID } from 'multiformats/cid'
 import { sha256 } from 'multiformats/hashes/sha2'
@@ -197,12 +198,12 @@ export class LimitTracker {
  * A custom matcher that tells us to listen on a particular relay
  */
 export const CircuitListen = fmt(
-  and(P2P.matchers[0], literal('p2p-circuit'))
+  and(P2P.matchers[0], code(CODE_P2P_CIRCUIT))
 )
 
 /**
  * A custom matcher that tells us to discover available relays
  */
 export const CircuitSearch = fmt(
-  literal('p2p-circuit')
+  code(CODE_P2P_CIRCUIT)
 )