+++ packages/metrics-prometheus/CHANGELOG.md
@@ -81,6 +81,24 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [4.3.28](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.27...prometheus-metrics-v4.3.28) (2025-07-07)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.14 to ^6.4.15
+
+## [4.3.27](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.26...prometheus-metrics-v4.3.27) (2025-07-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.13 to ^6.4.14
+
 ## [4.3.26](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.25...prometheus-metrics-v4.3.26) (2025-06-25)