+++ packages/protocol-perf/CHANGELOG.md
@@ -82,6 +82,24 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [4.0.45](https://github.com/libp2p/js-libp2p/compare/perf-v4.0.44...perf-v4.0.45) (2025-07-07)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.14 to ^6.4.15
+
+## [4.0.44](https://github.com/libp2p/js-libp2p/compare/perf-v4.0.43...perf-v4.0.44) (2025-07-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.13 to ^6.4.14
+
 ## [4.0.43](https://github.com/libp2p/js-libp2p/compare/perf-v4.0.42...perf-v4.0.43) (2025-06-25)