+++ packages/pubsub-floodsub/CHANGELOG.md
@@ -98,6 +98,24 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [10.1.44](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.43...floodsub-v10.1.44) (2025-07-07)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.14 to ^6.4.15
+
+## [10.1.43](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.42...floodsub-v10.1.43) (2025-07-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.13 to ^6.4.14
+
 ## [10.1.42](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.41...floodsub-v10.1.42) (2025-06-25)